Revert extraneous changes
