+++ config/manager/kustomization.yaml
@@ -21,5 +21,5 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
 - name: controller
-  newName: quay.io/jmatsuok/cryostat-operator
+  newName: quay.io/cryostat/cryostat-operator
   newTag: 4.2.0-dev
