+++ config/manifests/bases/cryostat-operator.clusterserviceversion.yaml
@@ -179,9 +179,9 @@ spec:
         path: declarativeCredentials[0].secretName
         x-descriptors:
         - urn:alm:descriptor:io.kubernetes:Secret
-      - description: Enable Cryostat audit logging. New Cryostat CRs default to true,
+      - description: 'Enable Cryostat audit logging. New Cryostat CRs default to true,
           but existing CRs created before this field was introduced remain unchanged
-          until this property is explicitly set.
+          until this property is explicitly set.'
         displayName: Enable Audit Logging
         path: enableAudit
         x-descriptors:
