+++ internal/main.go
@@ -243,13 +243,6 @@ func main() {
 		setupLog.Error(err, "unable to create webhook", "webhook", "Pod")
 		os.Exit(1)
 	}
-	agentDeploymentWebhook := agent.NewAgentDeploymentWebhook(&agent.AgentDeploymentWebhookConfig{
-		FIPSEnabled: fipsEnabled,
-	})
-	if err = agentDeploymentWebhook.SetupWebhookWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create webhook", "webhook", "Pod")
-		os.Exit(1)
-	}
 	// +kubebuilder:scaffold:builder
 
 	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
