+++ internal/webhooks/agent/pod_webhook.go
@@ -19,6 +19,7 @@ import (
 
 	operatorv1beta2 "github.com/cryostatio/cryostat-operator/api/v1beta2"
 	"github.com/cryostatio/cryostat-operator/internal/controllers/common"
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
@@ -28,6 +29,7 @@ import (
 
 // podWebhookLog is for logging in this package.
 var podWebhookLog = logf.Log.WithName("pod-webhook")
+var deploymentWebhookLog = logf.Log.WithName("deployment-webhook")
 
 // Environment variable to override the agent init container image
 const agentInitImageTagEnv = "RELATED_IMAGE_AGENT_INIT"
@@ -65,6 +67,17 @@ func (r *agentWebhook) SetupWebhookWithManager(mgr ctrl.Manager) error {
 		return err
 	}
 
+	deploymentWebhook := admission.WithCustomDefaulter(mgr.GetScheme(), &appsv1.Deployment{}, &deploymentMutator{
+		client: mgr.GetClient(),
+		config: r.AgentWebhookConfig,
+		log:    &deploymentWebhookLog,
+		gvk:    &gvk,
+		ReconcilerTLS: common.NewReconcilerTLS(&common.ReconcilerTLSConfig{
+			Client: mgr.GetClient(),
+			OS:     r.OSUtils,
+		}),
+	}).WithRecoverPanic(true)
+
 	webhook := admission.WithCustomDefaulter(mgr.GetScheme(), &corev1.Pod{}, &podMutator{
 		client: mgr.GetClient(),
 		config: r.AgentWebhookConfig,
@@ -76,6 +89,9 @@ func (r *agentWebhook) SetupWebhookWithManager(mgr ctrl.Manager) error {
 		}),
 	}).WithRecoverPanic(true)
 	// Modify the webhook to never deny the pod from being admitted
+	deploymentWebhook.Handler = allowAllRequests(deploymentWebhook.Handler)
+	mgr.GetWebhookServer().Register("/mutate--v1-deployment", deploymentWebhook)
+
 	webhook.Handler = allowAllRequests(webhook.Handler)
 	mgr.GetWebhookServer().Register("/mutate--v1-pod", webhook)
 	return nil
